refactor: organize skill guidance modules
