+++ packages/ai/src/orchestration/orchestrate-standalone.ts
@@ -1,7 +1,8 @@
 import { getEnabledMcpServerCount } from "@notra/ai/integrations/mcp-tool-index";
 import { createModel } from "@notra/ai/model";
 import { getStandaloneChatPrompt } from "@notra/ai/prompts/standalone-chat";
-import { listSkillCatalog } from "@notra/ai/skills/service";
+import { STANDALONE_SKILL_CATALOG_LIMIT } from "@notra/ai/skills/constants";
+import { listSkillCatalog } from "@notra/ai/skills/functions/service";
 import { createLazyMcpRuntime } from "@notra/ai/tools/mcp-lazy";
 import type {
   AutoThinkingLevel,
@@ -285,7 +286,12 @@ export async function orchestrateStandaloneChat(
 }
 
 async function getStandaloneSkillSummaries(organizationId: string) {
-  return (await listSkillCatalog({ organizationId }, { limit: 30 })).skills;
+  return (
+    await listSkillCatalog(
+      { organizationId },
+      { limit: STANDALONE_SKILL_CATALOG_LIMIT }
+    )
+  ).skills;
 }
 
 function createStandaloneToolProvisioningRuntime({
