+++ packages/ai/src/skills/constants.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_SKILL_CATALOG_LIMIT = 20;
+export const STANDALONE_SKILL_CATALOG_LIMIT = 30;
+export const WHITESPACE_REGEX = /\s+/g;
